Read note before changing
People were complaining about the colour of the delete message button being too dark on the menu when you right click your message. It was set to var(--danger-color) Rather than forcing people to change their danger color for everything I set it to what it originally was. Feel free to change it back, but if you do you should tell people it's there --danger-color